Set up a menu for possible actions instead of using a 'pop-out' menu on the nodes themselves.
Added playwright for e2e testing framework and formally added typescript

Added the unit test cases from the beatiful-react-diagrams package

Added yarn as a package manager

Allowed slight movement in the nodes to count as a click.
The boundary of this movement is defined by clickSensitivity, which should
be hoisted up the component tree later.