Fixed bug within noCycles predicate. noCycles can now be used to validate links and will ensure that there are no cycles present in the produced schema.
Coerced the events returned by useDrag to match React synthetic events.
This is so that they can be used with the generic ClickHandler defined
in beautiful-react-diagrams/shared/Types.ts

Refactored onNodeSelect to make it more readable and to allow:
When a user doesn't press shift and clicks a node, within multiple
selected nodes, they expect the clicked node to remain selected.

Fixed a bug where selecting a nodes children deselected the node.